Update to build in dll
