+++ Ifnity/CMakeSettings.json
@@ -13,7 +13,7 @@
       "variables": [
         {
           "name": "BUILD_SHARED_IFNITY",
-          "value": "False",
+          "value": "True",
           "type": "BOOL"
         }
       ]
